+++ tests/files/match-default-export-name/id.js
@@ -1,0 +1,3 @@
+const myConstant = 42
+
+export default myConstant